feat: 로직 분리, 중심기반 패칭버튼
